+++ src/models/cql2/queryable.js
@@ -1,10 +1,10 @@
 import { formatKey } from "@radiantearth/stac-fields/helper";
 import i18n from '../../i18n.js';
-import { CqlEqual, CqlGreaterThan, CqlGreaterThanEqual, CqlLessThan, CqlLessThanEqual, CqlNotEqual } from "./operators/comparison";
-import { CqlLike } from "./operators/advanced";
+import { CqlEqual, CqlGreaterThan, CqlGreaterThanEqual, CqlLessThan, CqlLessThanEqual, CqlNotEqual, CqlBetween, CqlLike } from "./operators/comparison";
+import { CqlIn, CqlArrayOverlaps, CqlArrayContains, CqlArrayEquals, CqlArrayContainedBy } from "./operators/array";
+import { isObject } from "stac-js/src/utils.js";
 
 export default class Queryable {
-
   constructor(id, schema) {
     this.id = id;
     this.schema = schema;
@@ -25,7 +25,7 @@ export default class Queryable {
   }
 
   get supported() {
-    return this.isText || this.isNumeric || this.isBoolean;
+    return this.isText || this.isNumeric || this.isBoolean || this.isArray;
   }
 
   is(type) {
@@ -60,6 +60,27 @@ export default class Queryable {
     return this.isDate || this.isDateTime;
   }
 
+  get isArray() {
+    return this.is('array');
+  }
+
+  get arrayItems() {
+    // Detect the type(s) of the items in the array based on the JSON Schema.
+    // JSON Schema allows a variety of ways to specify both "items" and "types".
+    // We cater only for the most common cases here, which are:
+    // items: object
+    // type: omitted, array of strings, a single string
+    if (this.isArray && isObject(this.schema.items)) {
+      if (typeof this.schema.items.type === 'string') {
+        return [this.schema.items.type];
+      }
+      else if (Array.isArray(this.schema.items.type)) {
+        return this.schema.items.type;
+      }
+    }
+    return [];
+  }
+
   get defaultValue() {
     if (typeof this.schema.default !== 'undefined') {
       return this.schema.default;
@@ -72,7 +93,7 @@ export default class Queryable {
     }
     else if (this.isNumeric) {
       if (typeof this.schema.minimum !== 'undefined') {
-       return this.schema.minimum;
+        return this.schema.minimum;
       }
       return 0;
     }
@@ -81,6 +102,9 @@ export default class Queryable {
     }
     else if (this.isBoolean) {
       return false;
+    }
+    else if (this.isArray) {
+      return [];
     }
     return null;
   }
@@ -96,8 +120,9 @@ export default class Queryable {
   }
 
   getOperators(cql) {
-    let ops = [];
-    if (!this.isDateTime) {
+    const ops = [];
+
+    if (!this.isDateTime && !this.isArray) {
       // Although it is supported, comparing specific instances in time doesn't give predictable results.
       // For example 2020-01-01T00:00:00Z is not equal to 2020-01-01T00:00:00.001Z and you don't know the granularity
       // of the datetimes in the database. In the end you usually don't get what you are looking for.
@@ -105,15 +130,30 @@ export default class Queryable {
       ops.push(CqlEqual);
       ops.push(CqlNotEqual);
     }
+
     if (this.isNumeric || this.isTemporal) {
       ops.push(CqlLessThan);
       ops.push(CqlLessThanEqual);
       ops.push(CqlGreaterThan);
       ops.push(CqlGreaterThanEqual);
     }
-    else if (this.isText && cql.advancedComparison) {
-      ops.push(CqlLike);
+    if (this.isNumeric && cql.advancedComparison) {
+      ops.push(CqlBetween);
+      ops.push(CqlIn);
     }
+    if (this.isText && cql.advancedComparison) {
+      ops.push(CqlLike);
+      ops.push(CqlIn);
+    }
+
+    // Array operators for array-type queryables
+    if (this.isArray && cql.arrayOperators) {
+      ops.push(CqlArrayOverlaps);
+      ops.push(CqlArrayContains);
+      ops.push(CqlArrayEquals);
+      ops.push(CqlArrayContainedBy);
+    }
+
     return ops;
   }
 
@@ -124,5 +164,4 @@ export default class Queryable {
   toJSON() {
     return { property: this.id };
   }
-
 }