+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - PlayWright tests
 - Allow manually entering bounding boxes for search
 - Generate code examples for Global Item Search
+- Support negating CQL2 filters (globally and per filter)
+- Support CQL2 Advanced Comparison Operators
+- Support CQL2 Array Functions
 
 ### Changed
 
@@ -42,6 +45,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Handle state of downloads better and confirm leaving the page when downloading
 - Better error on request to the /collections or .../items endpoints
 - Collection list on Global Item Search was empty in certain situations
+- Show an error message when no operator is supported for a queryable
+- Don't show an "unsupported" error when only Collection Search is supported by the API
 
 ## [4.0.1] - 2026-02-11